chore: :recycle: review corrections

+++ apps/server/src/resources/zone/router.ts
@@ -2,6 +2,7 @@ import { serverInstance } from '@/app.js'
 import { addReqLogs } from '@/utils/logger.js'
 import { createZone, deleteZone, listZones, updateZone } from './business.js'
 import { zoneContract } from '@cpn-console/shared'
+import { assertIsAdmin } from '@/utils/controller.js'
 
 export const zoneRouter = () => serverInstance.router(zoneContract, {
   listZones: async ({ request: req }) => {
@@ -15,6 +16,7 @@ export const zoneRouter = () => serverInstance.router(zoneContract, {
   },
 
   createZone: async ({ request: req, body: data }) => {
+    assertIsAdmin(req.session.user)
     const zone = await createZone(data)
 
     addReqLogs({ req, message: 'Zone créée avec succès', infos: { zoneId: zone.id } })
@@ -25,6 +27,7 @@ export const zoneRouter = () => serverInstance.router(zoneContract, {
   },
 
   updateZone: async ({ request: req, params, body: data }) => {
+    assertIsAdmin(req.session.user)
     const zoneId = params.zoneId
 
     const zone = await updateZone(zoneId, data)
@@ -36,6 +39,7 @@ export const zoneRouter = () => serverInstance.router(zoneContract, {
   },
 
   deleteZone: async ({ request: req, params }) => {
+    assertIsAdmin(req.session.user)
     const zoneId = params.zoneId
 
     await deleteZone(zoneId)
